Cleaning and refactoring

+++ src/main/java/com/mr/egzamator/respository/MarkRepository.java
@@ -2,8 +2,11 @@ package com.mr.egzamator.respository;
 
 import com.mr.egzamator.model.Mark;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
+import java.util.Set;
+
 @Repository
-public interface MarkRepository extends JpaRepository<Mark,Integer> {
+public interface MarkRepository extends JpaRepository<Mark,Long> {
 }
